API

Field Types

Individual Fields

Field Algebra

Propagator

Wick Contraction

Field Types

Individual Fields

╭───────────────────────── Types hierarchy ───╮
│                                             │
│  KeldyshContraction.QField                  │
│ ━━━━━━━━━━━━━━━━━━━━━━━━━━━                 │
│             │                               │
│             ├── KeldyshContraction.QTerm    │
│             └── KeldyshContraction.QSym     │
│                 ├── Create                  │
│                 └── Destroy                 │
│                                             │
╰─────────────────────────────────────────────╯

KeldyshContraction.QField — Type

QField

Abstract type representing any expression involving Fields.

KeldyshContraction.QSym — Type

QSym <: QField

Abstract type representing fundamental Field types.

KeldyshContraction.Create — Type

Create <: QSym

Bosonic field representing the quantum field creation operator.

KeldyshContraction.Destroy — Type

Destroy <: QSym

Bosonic field representing the quantum field annihilation operator.

Field Properties

The field properties are determined by the Enum objects:

KeldyshContraction.KeldyshContour — Type

KeldyshContour `Quantum` `Classical`

Keldysh contour enum for the Keldysh quantum field. The Keldysh contour is used to distinguish the field on quantum and classical contour.

KeldyshContraction.Regularisation — Type

Regularisation `Plus` `Zero` `Minus`

Regularisation enum for the Keldysh quantum field. The regularisation is used to perform tadpole regularisation during the Wick contraction.

KeldyshContraction.Position — Type

Position `In` `Out` `Bulk`

Position enum for the Keldysh quantum field. The position is used to determine the coordinate of the field.

Field Constructors

It is expected to create the fields using the Create and Destroy constructors, together with the @qfields macro.

KeldyshContraction.@qfields — Macro

@qfields

Convenience macro for the construction of fields.

Examples

julia> using KeldyshContraction: Classical, Quantum

julia> @qnumbers ψ::Destroy(Classical)
(ψ,)

The created fields are callable to change a property of the individual fields:

using KeldyshContraction
using KeldyshContraction: In, Classical, position

@qfields ϕ::Destroy(Classical)

position(ϕ(In))

In::Position = 1

Field Algebra

╭──────────────────────────── Types hierarchy ───╮
│                                                │
│  KeldyshContraction.QField                     │
│ ━━━━━━━━━━━━━━━━━━━━━━━━━━━                    │
│             │                                  │
│             ├── KeldyshContraction.QSym        │
│             └── KeldyshContraction.QTerm       │
│                 ├── KeldyshContraction.QAdd    │
│                 └── KeldyshContraction.QMul    │
│                                                │
╰────────────────────────────────────────────────╯

KeldyshContraction.QTerm — Type

QTerm <: QField

Abstract type representing noncommutative expressions.

KeldyshContraction.QMul — Type

QMul <: QTerm

Represent a multiplication involving QSym types.

Fields:

arg_c: The commutative prefactor.

args_nc: A vector containing all QSym types.

KeldyshContraction.QAdd — Type

QAdd <: QTerm

Represent an addition involving QField and other types.

Propagator

KeldyshContraction.Propagator — Type

Propagator <: Number

Symbolic number representing the Propagator of two fields ϕ and ψ. By convention a propagator is shown as G(ϕ, ψ) with ψ the incoming Create field and ϕ the outgoing Destroy field.

See also: propagator

KeldyshContraction.PropagatorType — Type

PropagatorType `Keldysh`, `Advanced`, `Retarded`

The type of propagator taken of two fields with the x-y Contour where x is the Contour of the Destroy field and y the contour of the Create field.

Keldysh propagator of a Classical-Classical contour

Advanced propagator of a Quantum-Classical contour

Retarded propagator of a Classical-Quantum contour

The Quantum-Quantum propagator should always be zero.

KeldyshContraction.propagator — Function

propagator(ψ::QSym, ϕ::QSym)

Create the Keldysh two-point green's functions over two field ψ and ϕ. Here ψ is the incoming Create field and ψ the outgoing Destroy field. The created propagator will be of type SymbolicUtils.BasicSymbolic{Propagator{T}} where T is the type PropagatorType.

Wick Contraction

KeldyshContraction.wick_contraction — Function

wick_contraction(expr::Union{QAdd,QMul})

Compute all possible Wick contractions of quantum fields in the expression expr.

Wick contractions decompose products of quantum field operators into sums of products of propagators (two-point correlation functions). The rules of the contraction are:

Conservation (equal numbers of creation/annihilation operators)

Physicality (proper time ordering)

No quantum-quantum contractions

If the fields have a KeldyshContraction.Regularisation applied, the contractions are regularised. The KeldyshContraction.Regularisation property is set to zero after the reguralisation.

The function handles two types of inputs:

QAdd: Distributes the contraction over sums

QMul: Contracts products of fields into propagators

The function returns a new expression of propagators of type SymbolicUtils.Symbol.

```